distributed-message-queue-kafka-python/requirements.txt
fastapi
uvicorn[standard]
py-dotenv
kafka-python
aiokafka
sse_starlette